fix bug admin

+++ packages/web-admin/src/api/productsApi.ts
@@ -2,20 +2,22 @@ import api from "@/api/axiosInstance"; // bạn đã có file axiosInstance.ts n
 
 // 🟢 Lấy danh sách sản phẩm
 export const getAllProducts = async () => {
-    return api.get("/products");
+    const cacheBuster = `_=${new Date().getTime()}`;
+    // 👇 SỬA LẠI "pproducts" THÀNH "products" Ở ĐÂY
+    return api.get(`/admin/products?${cacheBuster}`);
 };
 
 // 🟢 Thêm sản phẩm
 export const createProduct = async (data: any) => {
-    return api.post("/products", data);
+    return api.post("/admin/products", data);
 };
 
 // 🟢 Cập nhật sản phẩm
 export const updateProduct = async (id: number, data: any) => {
-    return api.put(`/products/${id}`, data);
+    return api.put(`/admin/products/${id}`, data);
 };
 
 // 🟢 Xóa sản phẩm
 export const deleteProduct = async (id: number) => {
-    return api.delete(`/products/${id}`);
+    return api.delete(`/admin/products/${id}`);
 };
